Fixed DataLoader for one-time run, fixed N+1 issue, optimized requests in getPatient() method, and added new validation.

+++ src/main/java/com/faifly/model/doctor/DoctorOutputDto.java
@@ -8,5 +8,5 @@ import lombok.Data;
 public class DoctorOutputDto {
     private String firstName;
     private String lastName;
-    private Integer totalPatients;
+    private Long totalPatients;
 }
